Chatbot/Chatbot Deployment/requirements.txt
pandas
sklearn
python-telegram-bot==12.7
requests==2.22.0
pyfiglet==0.8.post1
numpy
joblib